+++ internal/delivery/http/request/product_request/paginate.go
@@ -1,6 +1,6 @@
 package product_request
 
 type GetProductWithPagination struct {
-	Page     *uint32 `json:"page" query:"page"`
-	PageSize *uint32 `json:"page_size" query:"page_size"`
+	Page     *uint64 `json:"page" query:"page"`
+	PageSize *uint64 `json:"page_size" query:"page_size"`
 } // @name GetProductWithPaginationRequest
